Calculate ranking metric values by contrast

Calculate ranking metric values to be used with fgsea2.

get_ranking(
  x,
  genes,
  metric = "-log10(P.Value)*sign(logFC)",
  contrast_column = "contrast"
)

Arguments

x

object of class data.frame. Typically results from limma_full.

genes

string; the name of a column in x. A ranking metric value will be calculated for each unique entry. Can be gene symbols, Entrez IDs, or Ensembl genes.

metric

string; specifies how the ranking metric will be calculated from x. Can be the name of a single column.

contrast_column

string; the name of a column in x. Entries in this column will be set as the names of the output list.

Value

A named vector (if no "contrast" column) or list of sorted ranking metric values for each contrast in x.

Details

The -log\(_{10}\)-transformed p-value signed by the fold-change is similar to using t-statistics, but provides better separation between the extreme values in the tails and those close to 0 that are not as interesting. The log2 fold-change should not be used as the ranking metric, as it does not capture variability in the measurements.