Run WGCNA Pipeline

Run the WGCNA pipeline used in the WAT manuscript.

run_WGCNA(
  object,
  power = 1:20,
  RsquaredCut = 0.9,
  module_prefix = "",
  merge_modules
)

Arguments

object

object of class MSnSet. exprs should be a matrix of log\(_2\)-transformed values. Count data is not accepted.

power

integer; optional soft power vector (length 1 or greater). If not specified, the lowest value of 1:20 that satisfies scale-free fit R\(^2\) at least equal to RsquaredCut will be used.

RsquaredCut

numeric; desired minimum scale-free fit R\(^2\). Ignored if power is provided.

module_prefix

character; what to append to the module numbers to create the "moduleID" column in fData(object).

merge_modules

logical; if not provided, user input is requested to determine whether highly-correlated (r=0.85) modules should be combined based on plots that the function creates. Must be provided if session is not interactive.

Value

Object of class MSnSet with modified pData and fData. The pData will include columns for each of the module eigenfeatures (prefixed with "ME" and followed by a color), while the fData will include columns "moduleColor" and "moduleID" that specify the modules to which each feature belongs. Features from the "grey" module should be ignored.

References

Zhang, B., & Horvath, S. (2005). A general framework for weighted gene co-expression network analysis. Statistical applications in genetics and molecular biology, 4, Article17. https://doi.org/10.2202/1544-6115.1128

Langfelder P and Horvath S, WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics 2008, 9:559 https://doi.org/10.1186/1471-2105-9-559

Peter Langfelder, Steve Horvath (2012). Fast R Functions for Robust Correlations and Hierarchical Clustering. Journal of Statistical Software, 46(11), 1--17. http://www.jstatsoft.org/v46/i11/

Horvath, S. (2011). Weighted Network Analysis. Springer New York. https://doi.org/10.1007/978-1-4419-8819-5

Langfelder P, Zhang B, Horvath wcfS (2016). dynamicTreeCut: Methods for Detection of Clusters in Hierarchical Clustering Dendrograms. R package version 1.63-1, https://CRAN.R-project.org/package=dynamicTreeCut.

Author

Tyler Sagendorf, Zhenxin Hou